backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
aiohttp==3.9.1
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
scipy==1.11.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1